Add prove PoQ benchmarks
